Fix Prevent URL Redirection
In file: webmentionAdmin.go, there is a function webmentionAdminAction that performs an HTTP redirect operation based upon data that comes from an external source without proper validation. This may allow an attacker to redirect users to a malicious site and trick the user allowing the attacker to steal the user's credentials. iCR suggested proper sanitization to prevent this Open Redirection attack. To learn more check CWE-601.